Corrected errors occuring due to pedantic check
@cvering and @tstorek, just for your information:
I am done with the physical modelling.
I will add the Users Guide and information section
at the weekend. However, I have to wait with the
pull request until the refrigerant library is added
to the development branch since its models are
required for the example models implemented here.

For #457