Solved debugging problem with Fluid Source model
The problem that every error in a model which contains the FluidSource
is associated with the FluidSource has been solved.
For that the EnergyBalance model and the EnthalpyPort model are changed
so the massflow is always positiv. (no more flow variable used)
This Fix is ok because in FastHvac no change of flow direction is
possible.