#563 Fixes all of Dymola's HTML code errors
This is needed to satisfy Travis CI.

Dymola doesn't handle the _span_ directives as demanded by tidylib.
Seems that Dymola forgets to close open directives if they are nested.
As in here for example:

`<span style=\"font-family: Courier New;\">0.5*<span style=\"color: #ff0000;\">max</span>`

Also the nesting of <i> and <span> was not done properly by Dymola
as in here:

```
Changed parameter name <i>
  <span style=\"font-family: Courier New;\">n_start</i> to
```

This seems to confuse tidylib.