WIP: Adds ThermalZoneMoistAir model extending from ThermalZone
This model is extended from the ThermalZone. It redeclares the Air
volume to the MixingVolumeMoistAir and sets the boolean parameter
use_moisture_balance to true. Thus the moisture balance is considered.

Still there is an error when checking. If setting up a simulation this
error does not occur. Has to be checked.
For #695