Travis: build with Python 3
Travis CI currently uses Python 2.7 and fails on `master`. The underlying problem is this:

```
  File "/home/travis/virtualenv/python2.7.15/lib/python2.7/site-packages/buildingspy/development/validator.py", line 84, in _getInfoRevisionsHTML

    with open(moFile, mode="r", encoding="utf-8") as f:

TypeError: 'encoding' is an invalid keyword argument for this function
```
https://travis-ci.org/github/RWTH-EBC/AixLib/jobs/734457254#L295

In Python 2.7, `open` does not accept an `encoding` parameter. In Python 3.8, it does.
https://docs.python.org/2.7/library/functions.html#open
https://docs.python.org/3.8/library/functions.html#open

Additionally, the call stack clearly shows that `open` is called from the BuildingsPy library.
The BuildingsPy library build instruction make it clear that Python 3 is required.
https://simulationresearch.lbl.gov/modelica/buildingspy/install.html

Python 3.9 is the latest and greatest Python version, but it was released less than a week ago.
Python 3.8 was released a year ago and still receives bugfixes (as opposed to older versions which only receive security updates), so that seems like a good compromise.